ADDED: Security handling for interview
- Added handling for changing tabs, if someone changes tab for more than 3 times, the user will get disqualified from interview and will be logged out from the portal
- Removed useless code from App.js file
